add authorization

+++ backend/routes/posts.js
@@ -5,7 +5,11 @@ const checkAuth = require("../middleware/auth");
 const router = express.Router();
 
 router.post("", checkAuth, (req,res,next) => {
-  post = new Post({ title: req.body.title, content: req.body.content });
+  post = new Post({ 
+    title: req.body.title, 
+    content: req.body.content,
+    creator: req.userData.userId 
+  });
   post.save().then(result => {
     res.status(201).json({
       message:"Post Added Successfully",
@@ -34,14 +38,29 @@ router.get("/:id", (req,res,next) => {
 })
 
 router.put("/:id", checkAuth, (req,res,next) => {
-  Post.updateOne({_id: req.params.id}, { title: req.body.title, content: req.body.content }).then(result => {
-    res.status(200).json({ message: "Update Sucessful" });
+  const post = ({
+    _id: req.body.id,
+    title: req.body.title,
+    content: req.body.content,
+    creator: req.userData.userId
+  });
+
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if(result.nModified > 0) {
+      res.status(200).json({ message: "Update Sucessful" });
+    }else {
+      res.status(401).json({ message: "User is not authorized to update this post" });
+    }
   });
 });
 
 router.delete("/:id", checkAuth, (req,res,next) => {
-  Post.deleteOne({_id:req.params.id}).then(result => {
-    return res.status(200).json({ mmesage: "Post Deleted" });
+  Post.deleteOne({_id:req.params.id, creator: req.userData.userId }).then(result => {
+    if(result.n > 0) {
+      return res.status(200).json({ mmesage: "Post Deleted" });
+    }else {
+      return res.status(401).json({ mmesage: "User is not authorized to delete this post" });
+    }
   });
 });
